Linear Regression with Problem data-set
Simple linear regression is used to model the linear relationship between one predictor variable and a continuous target variable. It attempts to find the relationship between the predictor and target variables so that we can predict a target value. For example, based on certain house features (predictors) such as number of bedrooms and total square feet, we can predict house prices (target)

Multiple linear regression uses multiple features to model a linear relationship with a target variable. It is used when we have more than one predictor variable.